move cli tests to npm integration
